% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_canopydensity.R
\docType{data}
\name{metrics_canopydensity}
\alias{metrics_canopydensity}
\alias{.metrics_canopydensity}
\title{Canopy density metrics}
\format{
An object of class \code{formula} of length 2.
}
\usage{
metrics_canopydensity(z, interval_count = 10, zmin = NA)

.metrics_canopydensity
}
\arguments{
\item{z}{Z coordinate of the point cloud}

\item{interval_count}{Number of intervals used to divide the point height distribution.}

\item{zmin}{Minimum height. If set, heights below are ignored in calculations.}
}
\value{
Cumulative percentage of points in each interval
}
\description{
Cumulative distribution of point heights in interval count.
}
\keyword{datasets}
