% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_dispersion.R
\name{metrics_dispersion}
\alias{metrics_dispersion}
\title{Dispersion metrics}
\usage{
metrics_dispersion(z, dz = 1, zmin = NA)
}
\arguments{
\item{z}{Z coordinate of the point cloud}

\item{dz}{layer thickness to use when calculating entropy and VCI.}

\item{zmin}{Minimum height. If set, heights below are ignored in calculations.}
}
\value{
Interquartile distance, mean absolute deviation (MAD) around the mean and median, canopy relief ratio
}
\description{
Metrics characterizing variation of point cloud heights.
}
\examples{
library(lidR)
library(lidRmetrics)
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las <- readLAS(LASfile, select = "*", filter = "-keep_random_fraction 0.5")

m1 <- cloud_metrics(las, ~metrics_dispersion(z = Z))

m2 <- grid_metrics(las, ~metrics_dispersion(z = Z), res = 40)
}
