% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_canopydensity.R
\name{metrics_fhd}
\alias{metrics_fhd}
\title{Foliage height diversity}
\usage{
metrics_fhd(z, dz = 1, zmin = 0)
}
\arguments{
\item{z}{Z coordinate of the point cloud}

\item{dz}{Interval layer thickness}

\item{zmin}{Minimum height. If set, heights below are ignored in calculations.}
}
\value{
Cumulative percentage of points in each interval
}
\description{
Foliage height diversity
}
