% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_HOME.R
\name{metrics_HOME}
\alias{metrics_HOME}
\title{Height of median energy (HOME)}
\usage{
metrics_HOME(z, i, zmin = NA)
}
\arguments{
\item{z}{Z coordinate of the point cloud}

\item{i}{Intensity}

\item{zmin}{Minimum height. If set, heights below are ignored in calculations.}
}
\value{
Height of median energy
}
\description{
The function provided here aims to mimic the HOME metric, which is a metric typically used with full-waveform lidar. 
HOME is calculated by identifying an elevation that splits the total intensity into two equal parts. Function is based
on a similar metric implemented in LAStools.
}
\examples{
library(lidR)
library(lidRmetrics)
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las <- readLAS(LASfile, select = "*", filter = "-keep_random_fraction 0.5")

m1 <- cloud_metrics(las, ~metrics_HOME(z = Z, i = Intensity))

m2 <- grid_metrics(las, ~metrics_HOME(z = Z, i = Intensity), res = 20)
}
