% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_interval.R
\name{metrics_interval}
\alias{metrics_interval}
\title{Interval metrics}
\usage{
metrics_interval(z, zintervals = c(0, 0.15, 2, 5, 10, 20, 30), zmin = NA)
}
\arguments{
\item{z}{Z coordinate of the point cloud}

\item{zintervals}{Height intervals}

\item{zmin}{Minimum height. If set, heights below are ignored in calculations.}
}
\value{
Percentage of points within each height interval
}
\description{
Percentage of points calculated for a set of horizontal layers.
}
