% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_lad.R
\name{metrics_lad}
\alias{metrics_lad}
\title{LAD metrics}
\usage{
metrics_lad(z, zmin = NA)
}
\arguments{
\item{z}{Z coordinate of the point cloud}

\item{zmin}{Minimum height. If set, heights below are ignored in calculations.}
}
\value{
min, max, mean, and cv, of the leaf area density profile.
}
\description{
Metrics based on the leaf area density. \code{lidR::LAD()} used to calculate the leaf area density.
}
\examples{
library(lidR)
library(lidRmetrics)
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las <- readLAS(LASfile, select = "*", filter = "-keep_random_fraction 0.5")

m1 <- cloud_metrics(las, ~metrics_lad(z = Z))

m2 <- grid_metrics(las, ~metrics_lad(z = Z), res = 40)
}
