% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_Lmoments.R
\name{metrics_Lmoments}
\alias{metrics_Lmoments}
\title{L-moments and L-moment ratios}
\usage{
metrics_Lmoments(z, zmin = NA)
}
\arguments{
\item{z}{Z coordinate of the point cloud}

\item{zmin}{Minimum height. If set, heights below are ignored in calculations.}
}
\value{
L-moments and L-moment ratios
}
\description{
Calculates L-moments and L-moment ratios of point cloud heihts.
}
