% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_percentiles.R
\name{metrics_percentiles}
\alias{metrics_percentiles}
\title{Percentiles}
\usage{
metrics_percentiles(z, zmin = NA)
}
\arguments{
\item{z}{Z coordinate of the point cloud}

\item{zmin}{Minimum height. If set, heights below are ignored in calculations.}
}
\value{
A set of height percentiles
}
\description{
Percentiles of point heights
}
