% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_sets.R
\name{metrics_set1}
\alias{metrics_set1}
\title{Pre-defined sets of metrics}
\usage{
metrics_set1(
  z,
  zmin = NA,
  threshold = c(2, 5),
  dz = 1,
  interval_count = 10,
  zintervals = c(0, 0.15, 2, 5, 10, 20, 30)
)
}
\arguments{
\item{z}{Z coordinate of the point cloud}

\item{zmin}{Minimum height. If set, heights below are ignored in calculations.}

\item{threshold}{Threshold height(s). See \link[=metrics_percabove]{metrics_percabove}.}

\item{dz}{layer thickness to use when calculating entropy and VCI.}

\item{interval_count}{Number of intervals used to divide the point height distribution. See \link[=metrics_canopydensity]{metrics_canopydensity}.}

\item{zintervals}{Height intervals. See \link[=metrics_interval]{metrics_interval}.}
}
\description{
Pre-defined sets of point cloud metrics composed of different metrics_* functions available in the package.
}
