% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_sets.R
\name{metrics_set1}
\alias{metrics_set1}
\alias{metrics_set2}
\alias{metrics_set3}
\title{Pre-defined sets of metrics}
\usage{
metrics_set1(
  z,
  zmin = NA,
  threshold = c(2, 5),
  dz = 1,
  interval_count = 10,
  zintervals = c(0, 0.15, 2, 5, 10, 20, 30)
)

metrics_set2(
  x,
  y,
  z,
  zmin = NA,
  threshold = c(2, 5),
  dz = 1,
  interval_count = 10,
  zintervals = c(0, 0.15, 2, 5, 10, 20, 30),
  pixel_size = 1,
  vox_size = 1
)

metrics_set3(
  x,
  y,
  z,
  i,
  ReturnNumber,
  NumberOfReturns,
  zmin = NA,
  threshold = c(2, 5),
  dz = 1,
  interval_count = 10,
  zintervals = c(0, 0.15, 2, 5, 10, 20, 30),
  pixel_size = 1,
  vox_size = 1
)
}
\arguments{
\item{zmin}{Minimum height. If set, heights below are ignored in calculations.}

\item{threshold}{Threshold height(s). See \link[=metrics_percabove]{metrics_percabove}.}

\item{dz}{layer thickness to use when calculating entropy and VCI.}

\item{interval_count}{Number of intervals used to divide the point height distribution. See \link[=metrics_canopydensity]{metrics_canopydensity}.}

\item{zintervals}{Height intervals. See \link[=metrics_interval]{metrics_interval}.}

\item{x, y, z}{X, Y, Z coordinates of the point cloud}

\item{pixel_size}{pixel size for calculating rumple index}

\item{vox_size}{voxel size for calculating voxel metrics}
}
\description{
Pre-defined sets of point cloud metrics composed of different metrics_* functions available in the package.
}
\details{
The three pre-defined sets of metrics include different number of the metrics_* functions and are provided 
to conventiently compute most of the metrics commonly used at the same time.
\describe{
  \item{\code{metrics_set1()}}{Included metrics: \code{metrics_basic()}, \code{metrics_percentiles()},  
  \code{metrics_percabove()}, \code{metrics_dispersion()},  \code{metrics_canopydensity()}, 
  \code{metrics_Lmoments()}, \code{metrics_lad()}, \code{metrics_interval()}.}
  \item{\code{metrics_set2()}}{Included metrics: All metrics in \code{metrics_set1()}, \code{metrics_rumple()},
  \code{metrics_voxels()}. }
  \item{\code{metrics_set3()}}{Included metrics: All metrics in \code{metrics_set2()}, \code{metrics_kde()}, 
   \code{metrics_echo()}, \code{metrics_HOME()}}.
}
Currently no set includes \code{metrics_texture()} as the function is considered experimental at this stage.
}
