% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_voxels.R
\name{metrics_voxels}
\alias{metrics_voxels}
\title{Voxel-based metrics}
\usage{
metrics_voxels(x, y, z, vox_size = 1, zmin = NA)
}
\arguments{
\item{x, y, z}{X, Y, Z coordinates of a point cloud to be converted into voxels}

\item{vox_size}{voxel size}

\item{zmin}{Minimum height. If set, heights below are ignored in calculations.}
}
\description{
A set of metrics calculated in a voxel space, designed to be used within the \code{pixel_metrics} or \code{cloud_metrics} function from the \code{lidR}
package. 
For convenience, a point cloud is converted to a voxel space on the fly, without the need of using additional processing steps. 
Note, that because of the additional computation required to convert a point cloud to voxels, calculating voxel-based metrics
is markedly slower than other metrics_* functions.
}
\details{
Calculated metrics include:
\itemize{
\item{\code{vn} count of filled voxels}
\item{\code{FRall} and \code{FRcanopy}: filled ratio. For \code{FRall} a ratio between the number of filled voxels and all voxels located in 
the maximum extent of the point cloud. For \code{FRcanopy} empty voxels above the canopy are excluded}
\item{metrics describing the vertical distribution of filled voxels: standard deviation (\code{vzsd}), 
coeficient of variation (\code{vzcv}), and vertical rumple (\code{vzrumple}).}
\item{Canopy volume classes based on Lefsky et al 1999 (see references), modified. A voxel representation of a forest stand
is divided into four classes including: open gap space, closed gap space, euphotic zone, and oligophotic zone.}

}
}
\examples{
library(lidR)
library(lidRmetrics)
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las <- readLAS(LASfile, select = "*", filter = "-keep_random_fraction 0.5")

m1 <- cloud_metrics(las, ~metrics_voxels(x = X, y = Y, z = Z, vox_size = 1))

m2 <- pixel_metrics(las, ~metrics_voxels(x = X, y = Y, z = Z, vox_size = 1), res = 20)
}
\references{
Lefsky, M. A., Cohen, W. B., Acker, S. A., Parker, G. G., Spies, T. A., & Harding, D. (1999). Lidar Remote Sensing of the Canopy Structure and Biophysical Properties of Douglas-Fir Western Hemlock Forests. Remote Sensing of Environment, 70(3), 339-361. doi:10.1016/S0034-4257(99)00052-8
}
