% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_canopydensity.R
\name{metrics_canopydensity}
\alias{metrics_canopydensity}
\title{Canopy density metrics}
\usage{
metrics_canopydensity(z, interval_count = 10, zmin = NA)
}
\arguments{
\item{z}{Z coordinate of the point cloud}

\item{interval_count}{Number of intervals used to divide the point height distribution.}

\item{zmin}{Minimum height. If set, heights below are ignored in calculations.}
}
\value{
Cumulative percentage of points in each interval
}
\description{
Cumulative distribution of point heights in interval count.
}
