% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_dispersion.R
\name{metrics_dispersion}
\alias{metrics_dispersion}
\title{Dispersion metrics}
\usage{
metrics_dispersion(z, dz = 1, zmin = NA)
}
\arguments{
\item{z}{Z coordinate of the point cloud (point heights)}

\item{dz}{Numeric. Layer thickness to use when calculating entropy and VCI.}

\item{zmin}{numeric. Minimum \code{z} value. If set, \code{z} values (heights) below are ignored in calculations.}
}
\value{
A list. Calculated metrics include:
\itemize{
\item \code{ziqr} interquartile distance
\item \code{zMADmean} mean absolute deviation (MAD) around the mean
\item \code{zMADmedian} mean absolute deviation (MAD) around the median
\item \code{CRR} canopy relief ratio
\item \code{zentropy} entropy
\item \code{VCI} vertical complexity index

}
}
\description{
Metrics characterizing the dispersion of point heights
}
\details{
When calculating \code{zentropy} and \code{VCI}, \code{z} values below 0 are removed.
}
\examples{
library(lidR)
library(lidRmetrics)
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las <- readLAS(LASfile, select = "*", filter = "-keep_random_fraction 0.5")

m1 <- cloud_metrics(las, ~metrics_dispersion(z = Z))

m2 <- pixel_metrics(las, ~metrics_dispersion(z = Z, dz = 2), res = 20)
}
