% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_echo.R
\name{metrics_echo}
\alias{metrics_echo}
\title{Number and proportion of points by echo types}
\usage{
metrics_echo(ReturnNumber, NumberOfReturns, z = NULL, zmin = NA)
}
\arguments{
\item{ReturnNumber}{return number}

\item{NumberOfReturns}{number of returns}

\item{z}{Z coordinate of the point cloud. Required only if \code{zmin} is used.}

\item{zmin}{Minimum height. If set, heights below are ignored in calculations.}
}
\value{
Number (n_first, n_intermediate, n_last, n_single, n_multiple) 
and percentage (p_first, p_intermediate, p_last, p_single, p_multiple) of returns by each echo type.
Ratio of First to Last, First to Intermediate, and Multiple to Single.
}
\description{
Based on the return number and number of returns, point cloud returns are classified into different echo types. Two different
classification routines are applied. Under the first classification routine returns are classified into First, Intermediate, and Last. 
Under the second classification routine returns are classified into Single or Multiple. Function then calculates point counts and proportions 
by each echo type. Ratios of First to Last, First to Intermediate, and Multiple to Single, are also calculated.
}
\examples{
library(lidR)
library(lidRmetrics)
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las <- readLAS(LASfile, select = "*", filter = "-keep_random_fraction 0.5")

m1 <- cloud_metrics(las, ~metrics_echo(ReturnNumber=ReturnNumber, NumberOfReturns=NumberOfReturns))

m2 <- pixel_metrics(las, ~metrics_echo(ReturnNumber=ReturnNumber, NumberOfReturns=NumberOfReturns), res = 20)
}
