% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_dispersion.R, R/metrics_echo.R
\docType{data}
\name{.metrics_dispersion}
\alias{.metrics_dispersion}
\alias{metrics_echo}
\alias{.metrics_echo}
\title{Number and proportion of points by echo types}
\format{
An object of class \code{formula} of length 2.

An object of class \code{formula} of length 2.
}
\usage{
.metrics_dispersion

metrics_echo(ReturnNumber, NumberOfReturns, z = NULL, zmin = NA)

.metrics_echo
}
\arguments{
\item{ReturnNumber}{return number}

\item{NumberOfReturns}{number of returns}

\item{z}{Z coordinate of the point cloud (point heights)}

\item{zmin}{numeric. Minimum \code{z} value. If set, \code{z} values (heights) below are ignored in calculations.}
}
\value{
A list. Calculated metrics include:
\itemize{
\item Number of first (\code{n_first}), intermediate (\code{n_intermediate}), last (\code{n_last}), single (\code{n_single}), and multiple  (\code{n_multiple}) returns
\item Proportion of first (\code{p_first}), intermediate (\code{p_intermediate}), last (\code{p_last}), single (\code{p_single}), and multiple  (\code{p_multiple}) returns
\item Ratio of Last to First (\code{ratio_last_first}), Intermediate to First (\code{ratio_intermediate_first}), and Multiple to Single (\code{ratio_multiple_single}). 
}
}
\description{
Based on the return number and number of returns, point cloud returns are classified into different echo types. Two different
classification routines are applied. Under the first classification routine returns are classified into First, Intermediate, and Last. 
Under the second classification routine returns are classified into Single or Multiple. Function then calculates point counts and proportions 
by each echo type. Ratios of Last to First, Intermediate to First, and Multiple to Single, are also calculated.
}
\examples{
library(lidR)
library(lidRmetrics)
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las <- readLAS(LASfile, select = "*", filter = "-keep_random_fraction 0.5")

m1 <- cloud_metrics(las, ~metrics_echo(ReturnNumber=ReturnNumber, NumberOfReturns=NumberOfReturns))

m2 <- pixel_metrics(las, ~metrics_echo(ReturnNumber=ReturnNumber, NumberOfReturns=NumberOfReturns), res = 20)
}
\keyword{datasets}
