% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_HOME.R
\docType{data}
\name{metrics_HOME}
\alias{metrics_HOME}
\alias{.metrics_HOME}
\title{Height of median energy (HOME)}
\format{
An object of class \code{formula} of length 2.
}
\usage{
metrics_HOME(z, i, zmin = NA)

.metrics_HOME
}
\arguments{
\item{z}{Z coordinate of the point cloud (point heights)}

\item{i}{Intensity}

\item{zmin}{numeric. Minimum \code{z} value. If set, \code{z} values (heights) below are ignored in calculations.}
}
\value{
Height of median energy
}
\description{
The function provided here aims to mimic the HOME metric, which is a metric typically used with full-waveform lidar. 
HOME is calculated by identifying an elevation that splits the total intensity into two equal parts. Function is based
on a similar metric implemented in LAStools.
}
\examples{
library(lidR)
library(lidRmetrics)
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las <- readLAS(LASfile, select = "*", filter = "-keep_random_fraction 0.5")

m1 <- cloud_metrics(las, ~metrics_HOME(z = Z, i = Intensity))

m2 <- grid_metrics(las, ~metrics_HOME(z = Z, i = Intensity), res = 20)
}
\keyword{datasets}
