% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_interval.R
\docType{data}
\name{metrics_interval}
\alias{metrics_interval}
\alias{.metrics_interval}
\title{Interval metrics}
\format{
An object of class \code{formula} of length 2.
}
\usage{
metrics_interval(z, zintervals = c(0, 0.15, 2, 5, 10, 20, 30), zmin = NA)

.metrics_interval
}
\arguments{
\item{z}{Z coordinate of the point cloud (point heights)}

\item{zintervals}{Numeric. Height intervals}

\item{zmin}{numeric. Minimum \code{z} value. If set, \code{z} values (heights) below are ignored in calculations.}
}
\value{
A list. Percentage of points within each height interval
}
\description{
Percentage of points calculated for a set of horizontal layers.
}
\examples{
library(lidR)
library(lidRmetrics)
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las <- readLAS(LASfile, select = "*", filter = "-keep_random_fraction 0.5")

m1 <- cloud_metrics(las, ~metrics_interval(z = Z))

m2 <- pixel_metrics(las, ~metrics_interval(z = Z, zintervals = c(0, 5, 10)), res = 20)
}
\keyword{datasets}
