% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_interval.R
\name{metrics_interval}
\alias{metrics_interval}
\title{Interval metrics}
\usage{
metrics_interval(z, zintervals = c(0, 0.15, 2, 5, 10, 20, 30), zmin = NA)
}
\arguments{
\item{z}{Z coordinate of the point cloud}

\item{zintervals}{Height intervals}

\item{zmin}{Minimum height. If set, heights below are ignored in calculations.}
}
\value{
Percentage of points within each height interval
}
\description{
Percentage of points calculated for a set of horizontal layers.
}
\examples{
library(lidR)
library(lidRmetrics)
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las <- readLAS(LASfile, select = "*", filter = "-keep_random_fraction 0.5")

m1 <- cloud_metrics(las, ~metrics_interval(z = Z))

m2 <- grid_metrics(las, ~metrics_interval(z = Z), res = 40)
}
