% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_kde.R
\name{metrics_kde}
\alias{metrics_kde}
\title{Kernel density estimation (KDE) metrics}
\usage{
metrics_kde(z, zmin = NA)
}
\arguments{
\item{z}{Z coordinate of the point cloud}

\item{zmin}{Minimum height. If set, heights below are ignored in calculations.}
}
\value{
Number of peaks, elevation and density value of the peak with largest density
}
\description{
Kernel density estimation applied to the distribution of point cloud elevation (Z). KDE allows to 
create a probability density function (using a Guassian kernel). The density function is then used to detect
peaks (function maxima). Based on similar metric available in Fusion (see references), modified.
}
\examples{
library(lidR)
library(lidRmetrics)
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las <- readLAS(LASfile, select = "*", filter = "-keep_random_fraction 0.5")

m1 <- cloud_metrics(las, ~metrics_kde(z = Z))

m2 <- grid_metrics(las, ~metrics_kde(z = Z), res = 20)
}
\references{
McGaughey, R.J., 2021. FUSION/LDV: Software for LIDAR Data Analysis and Visualization. http://forsys.cfr.washington.edu/software/fusion/FUSION_manual.pdf
}
