% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_lad.R
\name{metrics_lad}
\alias{metrics_lad}
\title{LAD metrics}
\usage{
metrics_lad(z, zmin = NA, dz = 1, k = 0.5, z0 = 2)
}
\arguments{
\item{z}{Z coordinate of the point cloud (point heights)}

\item{zmin}{numeric. Minimum \code{z} value. If set, \code{z} values (heights) below are ignored in calculations.}

\item{dz}{numeric. The thickness of the layers used (height bin)}

\item{k}{numeric. is the extinction coefficient}

\item{z0}{numeric. The bottom limit of the profile}
}
\value{
A list. \code{lad_min}, \code{lad_max}, \code{lad_mean}, \code{lad_cv}, and \code{lad_sum}.
}
\description{
Metrics based on the leaf area density. \code{lidR::LAD()} used to calculate the leaf area density.
}
\examples{
library(lidR)
library(lidRmetrics)
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las <- readLAS(LASfile, select = "*", filter = "-keep_random_fraction 0.5")

m1 <- cloud_metrics(las, ~metrics_lad(z = Z))

m2 <- pixel_metrics(las, ~metrics_lad(z = Z), res = 20)
}
