% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_percabove.R
\name{metrics_percabove}
\alias{metrics_percabove}
\title{Percent of points above threshold}
\usage{
metrics_percabove(z, threshold = c(2, 5), zmin = NA)
}
\arguments{
\item{z}{Z coordinate of the point cloud}

\item{threshold}{Threshold height(s).}

\item{zmin}{Minimum height. If set, heights below are ignored in calculations.}
}
\value{
Percent of points above thresholds
}
\description{
Calculates percentage of points above specified threshold heights (default = c(2, 5)) and mean height.
}
