% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_percentiles.R
\name{metrics_percentiles}
\alias{metrics_percentiles}
\title{Percentiles}
\usage{
metrics_percentiles(z, zmin = NA)
}
\arguments{
\item{z}{Z coordinate of the point cloud (point heights)}

\item{zmin}{numeric. Minimum \code{z} value. If set, \code{z} values (heights) below are ignored in calculations.}
}
\value{
A list. Set of height percentiles including: \code{zq1}, \code{zq5}, \code{zq10}, \code{zq15},
\code{zq20}, \code{zq25}, \code{zq30}, \code{zq35}, \code{zq40}, \code{zq45}, \code{zq50},
\code{zq55}, \code{zq60}, \code{zq65}, \code{zq70}, \code{zq75}, \code{zq80}, \code{zq85},
\code{zq90}, \code{zq95}, and \code{zq99}.
}
\description{
Percentiles of point heights
}
\examples{
library(lidR)
library(lidRmetrics)
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las <- readLAS(LASfile, select = "*", filter = "-keep_random_fraction 0.5")

m1 <- cloud_metrics(las, ~metrics_percentiles(z = Z))

m2 <- pixel_metrics(las, ~metrics_percentiles(z = Z, zmin = 2), res = 20)

}
