% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_rumple.R
\name{metrics_rumple}
\alias{metrics_rumple}
\title{Calculate rumple index}
\usage{
metrics_rumple(x, y, z, pixel_size, zmin = NA)
}
\arguments{
\item{x, y, z}{X, Y, Z coordinates of a point cloud}

\item{pixel_size}{pixel size}

\item{zmin}{Minimum height. If set, heights below are ignored in calculations.}
}
\value{
Same as in \code{lidR::rumple_index} - the calculated rumple index
}
\description{
A wrapper of the \code{lidR::rumple_index} function that allows to calculate rumple index without the need for CHM, and 
can be used directly in the e.g. \code{grid_metrics} function. The function combines the two required steps, i.e. creating a surface model, and calculating rumple index, into one.
Top surface is created using highest points within each pixel.
}
\examples{
library(lidR)
library(lidRmetrics)
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las <- readLAS(LASfile, select = "*", filter = "-keep_random_fraction 0.5")

m1 <- cloud_metrics(las, ~metrics_rumple(x = X, y = Y, z = Z, pixel_size = 1))

m2 <- grid_metrics(las, ~metrics_rumple(x = X, y = Y, z = Z, pixel_size = 1), res = 40)
}
